TinyLlamaChatApi/src/requirements.txt
flask_restful
llama-cpp-python